requirements.txt
discord.py==2.3.2
Requests==2.32.2
pydantic==2.7.1
SQLAlchemy=2.0.30